SignalR java client: Allow users to use a custom HttpClient implementation
Fixes #15404
