+++ src/SignalR/clients/java/signalr/src/main/java/com/microsoft/signalr/WebSocketOnClosedCallback.java
@@ -3,6 +3,6 @@
 
 package com.microsoft.signalr;
 
-interface WebSocketOnClosedCallback {
+public interface WebSocketOnClosedCallback {
     void invoke(Integer code, String reason);
 }
